fix: update shelter validation schema and created useAuthRoles hook to reduce complexity of code
