+++ src/hooks/index.ts
@@ -10,6 +10,7 @@ import { useSupplies } from './useSupplies';
 import { useViaCep } from './useViaCep';
 import { usePartners } from './usePartners';
 import { useGithubContributors } from './useGithubContributors';
+import { useAuthRoles } from './useAuthRoles';
 
 export {
   useShelters,
@@ -24,4 +25,5 @@ export {
   useViaCep,
   usePartners,
   useGithubContributors,
+  useAuthRoles,
 };
