+++ src/hooks/useAuthRoles/index.ts
@@ -1,0 +1,3 @@
+import { useAuthRoles } from './useAuthRoles';
+
+export { useAuthRoles };
