+++ src/pages/UpdateShelter/UpdateShelter.tsx
@@ -2,7 +2,6 @@ import { useContext, useEffect } from 'react';
 import { ChevronLeft, Loader } from 'lucide-react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { useFormik } from 'formik';
-import * as Yup from 'yup';
 import ReactSelect from 'react-select';
 
 import {
@@ -22,6 +21,8 @@ import { clearCache } from '@/api/cache';
 import { hardCodedRsCities } from '../CreateShelter/hardcodedCities';
 import { useDebouncedValue, useViaCep } from '@/hooks';
 import { cn } from '@/lib/utils';
+import { FullUpdateShelterSchema, UpdateShelterSchema } from './types';
+import { useAuthRoles } from '@/hooks/useAuthRoles/useAuthRoles';
 
 const UpdateShelter = () => {
   const navigate = useNavigate();
@@ -29,6 +30,7 @@ const UpdateShelter = () => {
   const { shelterId = '-1' } = params;
   const { data: shelter, loading } = useShelter(shelterId);
   const { session } = useContext(SessionContext);
+  const isAuthenticated = useAuthRoles('Staff');
 
   const {
     errors,
@@ -51,7 +53,7 @@ const UpdateShelter = () => {
       street: shelter.street ?? '',
       neighbourhood: shelter.neighbourhood ?? '',
       city: shelter.city ?? '',
-      streetNumber: shelter.streetNumber ?? null,
+      streetNumber: shelter.streetNumber,
       zipCode: shelter.zipCode ?? '',
       name: shelter.name,
     },
@@ -59,25 +61,11 @@ const UpdateShelter = () => {
     validateOnBlur: false,
     validateOnChange: false,
     validateOnMount: false,
-    validationSchema: Yup.object().shape({
-      shelteredPeople: Yup.number().nullable(),
-      petFriendly: Yup.bool().required('Este campo deve ser preenchido'),
-      verified: Yup.bool(),
-      address: Yup.string().nullable(),
-      capacity: Yup.string().nullable(),
-      pix: Yup.string().nullable(),
-      name: Yup.string(),
-      street: Yup.string().required('Este campo deve ser preenchido'),
-      neighbourhood: Yup.string().required('Este campo deve ser preenchido'),
-      city: Yup.string().required('Este campo deve ser preenchido'),
-      streetNumber: Yup.string()
-        .min(0, 'O valor mínimo para este campo é 1')
-        .required('Este campo deve ser preenchido'),
-      zipCode: Yup.string().nullable(),
-    }),
+    validationSchema: session ? FullUpdateShelterSchema : UpdateShelterSchema,
     onSubmit: async (values) => {
       try {
-        if (session) await ShelterServices.adminUpdate(shelterId, values);
+        if (isAuthenticated)
+          await ShelterServices.adminUpdate(shelterId, values);
         else await ShelterServices.update(shelterId, values);
         toast({
           title: 'Atualização feita com sucesso',
@@ -93,6 +81,7 @@ const UpdateShelter = () => {
       }
     },
   });
+
   const debouncedZipcode = useDebouncedValue(
     touched?.zipCode ? values?.zipCode ?? '' : '',
     500
@@ -107,8 +96,6 @@ const UpdateShelter = () => {
     if (cepData.logradouro) setFieldValue('street', cepData.logradouro);
     if (cepData.bairro) setFieldValue('neighbourhood', cepData.bairro);
     if (cepData.localidade) setFieldValue('city', cepData.localidade);
-
-    setErrors({});
   }, [cepData, setFieldValue, setErrors]);
 
   if (loading) return <LoadingScreen />;
